fix: Memoized the Form Control Component reducing unwanted re-rendering (#10553)
* fix: reduced unnecessary re-rendering of form components

* fix: memoized form control and moved config errors to selector logic
