+++ app/client/src/selectors/formSelectors.ts
@@ -1,9 +1,13 @@
 import { getFormValues, isValid, getFormInitialValues } from "redux-form";
 import { AppState } from "reducers";
 import { ActionData } from "reducers/entityReducers/actionsReducer";
-import { UIComponentTypes } from "../api/PluginApi";
-import { Plugin } from "api/PluginApi";
 import { FormEvaluationState } from "reducers/evaluationReducers/formEvaluationReducer";
+import { createSelector } from "reselect";
+import _ from "lodash";
+import { getDataTree } from "./dataTreeSelectors";
+import { DataTree } from "entities/DataTree/dataTreeFactory";
+import { Action } from "entities/Action";
+import { EvaluationError } from "utils/DynamicBindingUtils";
 
 type GetFormData = (
   state: AppState,
@@ -26,19 +30,42 @@ export const getApiName = (state: AppState, id: string) => {
 export const getFormEvaluationState = (state: AppState): FormEvaluationState =>
   state.evaluations.formEvaluation;
 
-// Selector to get UIComponent type from the redux state
-export const getUIComponent = (pluginId: string, allPlugins: Plugin[]) => {
-  let uiComponent = UIComponentTypes.DbEditorForm;
+type ConfigErrorProps = { configProperty: string; formName: string };
 
-  if (!!pluginId) {
-    // Adding uiComponent field to switch form type to UQI or allow for backward compatibility
-    const plugin = allPlugins.find((plugin: Plugin) =>
-      !!pluginId ? plugin.id === pluginId : false,
-    );
-    // Defaults to old value, new value can be DBEditorForm or UQIDBEditorForm
-    if (plugin) {
-      uiComponent = plugin.uiComponent;
+export const getConfigErrors = createSelector(
+  getDataTree,
+  (state: AppState, props: ConfigErrorProps) =>
+    getFormValues(props.formName)(state),
+  (_: AppState, props: ConfigErrorProps) => props.configProperty,
+  (dataTree: DataTree, formValues: Partial<Action>, configProperty: string) => {
+    // action that corresponds to this form control
+    let action: any;
+    let configErrors: EvaluationError[] = [];
+
+    // if form value exists, use the name of the form(which is the action's name) to get the action details
+    // from the data tree, then store it in the action variable
+    if (formValues && formValues.name) {
+      if (formValues.name in dataTree) {
+        // get action details from data tree
+        action = dataTree[formValues.name];
+
+        // extract the error object from the action's details object.
+        const actionError = action && action?.__evaluation__?.errors;
+
+        // get the configProperty for this form control and format it to resemble the format used in the action details errors object.
+        const formattedConfig = _.replace(
+          configProperty,
+          "actionConfiguration",
+          "config",
+        );
+
+        // grab the errors specific to this configProperty and store it in configErrors.
+        if (actionError && formattedConfig in actionError) {
+          configErrors = actionError[formattedConfig];
+        }
+      }
     }
-  }
-  return uiComponent;
-};
+
+    return configErrors;
+  },
+);
